feat: add dropout stats to monitoring
