+++ designer_v2/lib/features/monitor/study_monitor_table.dart
@@ -1,106 +1,8 @@
-import 'dart:convert';
-import 'dart:math';
-
-import 'package:equatable/equatable.dart';
 import 'package:flutter/material.dart';
 import 'package:intl/intl.dart';
-import 'package:studyu_core/core.dart';
 import 'package:studyu_designer_v2/common_views/standard_table.dart';
+import 'package:studyu_designer_v2/domain/study_monitoring.dart';
 import 'package:studyu_designer_v2/localization/app_translation.dart';
-
-class StudyMonitorItem extends Equatable {
-  final String participantId;
-  final String? inviteCode;
-  final DateTime enrolledAt;
-  final DateTime lastActivityAt;
-  final int currentDayOfStudy;
-  final int studyDurationInDays;
-  final int completedInterventions;
-  final int missedInterventions;
-  final int completedSurveys;
-  final int missedSurveys;
-  final String rawData;
-
-  const StudyMonitorItem({
-    required this.participantId,
-    required this.inviteCode,
-    required this.enrolledAt,
-    required this.lastActivityAt,
-    required this.currentDayOfStudy,
-    required this.studyDurationInDays,
-    required this.completedInterventions,
-    required this.missedInterventions,
-    required this.completedSurveys,
-    required this.missedSurveys,
-    required this.rawData,
-  });
-
-  static List<StudyMonitorItem> fromStudy(Study study) {
-    final List<StudyMonitorItem> items = [];
-
-    final participants = study.participants ?? [];
-    final participantsProgress = study.participantsProgress ?? [];
-
-    participants.sort((a, b) => a.startedAt!.compareTo(b.startedAt!)); // ascending
-
-    for (final participant in participants) {
-      final progresses = participantsProgress.where((progress) => progress.subjectId == participant.id).toList();
-      progresses.sort((b, a) => a.completedAt!.compareTo(b.completedAt!)); // descending
-      final interventionOrder = study.schedule.generateInterventionIdsInOrder(participant.selectedInterventionIds);
-      final lastActivityAt = progresses.isNotEmpty ? progresses.first.completedAt! : participant.startedAt!;
-      final studyDurationInDays = study.schedule.length;
-      final currentDayOfStudy =
-          min(studyDurationInDays, DateTime.now().toUtc().difference(participant.startedAt!).inDays);
-      final daysInBaseline = study.schedule.includeBaseline ? study.schedule.phaseDuration : 0;
-
-      int totalInterventions = 0;
-      for (int day = 0; day < currentDayOfStudy; day++) {
-        if (day < daysInBaseline) {
-          continue;
-        }
-
-        final interventionIdForThisPhase = interventionOrder[day ~/ study.schedule.phaseDuration];
-        final interventionForThisPhase = study.interventions.firstWhere((i) => i.id == interventionIdForThisPhase);
-        totalInterventions += interventionForThisPhase.tasks.length;
-      }
-
-      final totalSurveys = currentDayOfStudy * study.observations.length;
-
-      final completedInterventions = progresses.where((p) => p.resultType == "bool").toList();
-      final completedSurveys = progresses.where((p) => p.resultType != "bool").toList();
-
-      final missedInterventions = totalInterventions - completedInterventions.length;
-      final missedSurveys = totalSurveys - completedSurveys.length;
-
-      final rawDataDict = {
-        "participant": participant.toJson(),
-        "interventions": completedInterventions.map((p) => p.toJson()).toList(),
-        "surveys": completedSurveys.map((p) => p.toJson()).toList(),
-      };
-
-      final rawData = const JsonEncoder.withIndent('    ').convert(rawDataDict);
-
-      items.add(StudyMonitorItem(
-        participantId: participant.id,
-        inviteCode: participant.inviteCode,
-        enrolledAt: participant.startedAt!,
-        lastActivityAt: lastActivityAt,
-        currentDayOfStudy: currentDayOfStudy,
-        studyDurationInDays: studyDurationInDays,
-        completedInterventions: completedInterventions.length,
-        missedInterventions: missedInterventions,
-        completedSurveys: completedSurveys.length,
-        missedSurveys: missedSurveys,
-        rawData: rawData,
-      ));
-    }
-
-    return items;
-  }
-
-  @override
-  List<Object?> get props => [participantId];
-}
 
 class StudyMonitorTable extends StatelessWidget {
   final List<StudyMonitorItem> studyMonitorItems;
