+++ designer_v2/lib/localization/app_de.arb
@@ -490,9 +490,11 @@
   "monitoring_no_participants_description": "Sobald Teilnehmer in Ihrer Studie eingeschrieben sind, können Sie hier deren Fortschritt überwachen und ihre Daten einsehen.",
   "monitoring_participants_title": "Teilnehmer",
   "monitoring_active": "Aktiv",
-  "monitoring_active_tooltip": "Anzahl der Teilnehmer, die in den letzten 3 Tagen aktiv waren",
-  "monitoring_ended": "Beendet",
-  "monitoring_ended_tooltip": "Anzahl der Teilnehmer, die das Ende der Studie erreicht haben",
+  "monitoring_active_tooltip": "Anzahl der Teilnehmer, die derzeit in der Studie sind",
+  "monitoring_dropout": "Dropout",
+  "monitoring_dropout_tooltip": "Anzahl der Teilnehmer, die die Studie vorzeitig abgebrochen haben",
+  "monitoring_completed": "Abgeschlossen",
+  "monitoring_completed_tooltip": "Anzahl der Teilnehmer, die das Ende der Studie erreicht haben",
   "monitoring_table_column_participant_id": "Teilnehmer-ID",
   "monitoring_table_column_invite_code": "Einladungscode",
   "monitoring_table_column_enrolled": "Eingeschrieben",
