+++ designer_v2/lib/localization/app_en.arb
@@ -538,9 +538,11 @@
   "monitoring_no_participants_description": "Once participants have enrolled in your study, you can monitor their progress and view their data here.",
   "monitoring_participants_title": "Participants",
   "monitoring_active": "Active",
-  "monitoring_active_tooltip": "Number of participants that were active in the last 3 days",
-  "monitoring_ended": "Ended",
-  "monitoring_ended_tooltip": "Number of participants who reached the end of the study",
+  "monitoring_active_tooltip": "Number of participants who are currently in the study",
+  "monitoring_dropout": "Dropout",
+  "monitoring_dropout_tooltip": "Number of participants who have dropped out before the end of the study",
+  "monitoring_completed": "Completed",
+  "monitoring_completed_tooltip": "Number of participants who reached the end of the study",
   "monitoring_table_column_participant_id": "Participant ID",
   "monitoring_table_column_invite_code": "Invite Code",
   "monitoring_table_column_enrolled": "Enrolled",
